Resolve remaining race conditions and unify type normalization

+++ tests/qmlapicheck.sh
@@ -28,18 +28,19 @@ CPP="Ubuntu.Components Ubuntu.Components.ListItems Ubuntu.Components.Styles Ubun
 
 ERRORS=0
 echo Dumping QML API of C++ components
-rm $BUILD_DIR/components.api.new
+test -s $BUILD_DIR/components.api.new && rm $BUILD_DIR/components.api.new
     echo "Processing $i"
     # Silence spam on stderr due to fonts
     # https://bugs.launchpad.net/ubuntu-ui-toolkit/+bug/1256999
     # https://bugreports.qt-project.org/browse/QTBUG-36243
     env ALARM_BACKEND=memory $BUILD_DIR/tests/apicheck/apicheck \
-        --qml "$CPP" $BUILD_DIR/modules 1>> $BUILD_DIR/components.api.new
+        --qml "$CPP" $BUILD_DIR/modules 1>> $BUILD_DIR/components.api.new &&
+        echo Verifying the diff between existing and generated API
     test $? != 0 && echo Error: apicheck failed && ERRORS=1
-    echo Verifying the diff between existing and generated API
-# FIXME: Not clear why QTestRootObject shows up as a parent class
-test -s $BUILD_DIR/components.api.new && sed -r -i 's@QTestRootObject@QtObject@g' $BUILD_DIR/components.api.new && diff -Fqml -u $SRC_DIR/components.api $BUILD_DIR/components.api.new
-test $? != 0 && echo Error: Differences in API. Did you forget to update components.api? && ERRORS=1
+if [ "x$ERRORS" != "x1" ]; then
+    diff -F '[.0-9]' -u $SRC_DIR/components.api $BUILD_DIR/components.api.new
+    test $? != 0 && echo Error: Differences in API. Did you forget to update components.api? && ERRORS=1
+fi
 
 if [ "x$ERRORS" != "x1" ]; then
     echo API is all fine.
